Minor refactoring and cleanup of LINQ unit tests (#14)
