+++ Stormpath.SDK/Stormpath.SDK.Tests/Impl/Linq/SingleAsync_tests.cs
@@ -30,7 +30,7 @@ namespace Stormpath.SDK.Tests.Impl.Linq
     public class SingleAsync_tests : Linq_tests
     {
         [Fact]
-        public async Task SingleAsync_returns_one()
+        public async Task Returns_one_item()
         {
             var fakeDataStore = new FakeDataStore<IAccount>(new List<IAccount>()
                 {
@@ -44,7 +44,7 @@ namespace Stormpath.SDK.Tests.Impl.Linq
         }
 
         [Fact]
-        public void SingleAsync_throws_when_more_than_one_item_exists()
+        public void Throws_when_more_than_one_item_exists()
         {
             var fakeDataStore = new FakeDataStore<IAccount>(new List<IAccount>()
                 {
@@ -61,7 +61,7 @@ namespace Stormpath.SDK.Tests.Impl.Linq
         }
 
         [Fact]
-        public void SingleAsync_throws_when_no_items_exist()
+        public void Throws_when_no_items_exist()
         {
             var fakeDataStore = new FakeDataStore<IAccount>(Enumerable.Empty<IAccount>());
             var harness = CollectionTestHarness<IAccount>.Create<IAccount>(this.Href, fakeDataStore);
@@ -71,48 +71,6 @@ namespace Stormpath.SDK.Tests.Impl.Linq
             {
                 var jabba = await harness.Queryable.SingleAsync();
             });
-        }
-
-        [Fact]
-        public async Task SingleOrDefaultAsync_returns_one()
-        {
-            var fakeDataStore = new FakeDataStore<IAccount>(new List<IAccount>()
-                {
-                    FakeAccounts.HanSolo
-                });
-            var harness = CollectionTestHarness<IAccount>.Create<IAccount>(this.Href, fakeDataStore);
-
-            var han = await harness.Queryable.SingleOrDefaultAsync();
-
-            han.Surname.ShouldBe("Solo");
-        }
-
-        [Fact]
-        public void SingleOrDefaultAsync_throws_when_more_than_one_item_exists()
-        {
-            var fakeDataStore = new FakeDataStore<IAccount>(new List<IAccount>()
-                {
-                    FakeAccounts.HanSolo,
-                    FakeAccounts.LukeSkywalker
-                });
-            var harness = CollectionTestHarness<IAccount>.Create<IAccount>(this.Href, fakeDataStore);
-
-            // TODO This should be InvalidOperationException, but under Mono it throws NullReferenceException for some undetermined reason
-            Should.Throw<Exception>(async () =>
-            {
-                var han = await harness.Queryable.SingleOrDefaultAsync();
-            });
-        }
-
-        [Fact]
-        public async Task SingleOrDefaultAsync_returns_null_when_no_items_exist()
-        {
-            var fakeDataStore = new FakeDataStore<IAccount>(Enumerable.Empty<IAccount>());
-            var harness = CollectionTestHarness<IAccount>.Create<IAccount>(this.Href, fakeDataStore);
-
-            var notHan = await harness.Queryable.SingleOrDefaultAsync();
-
-            notHan.ShouldBeNull();
         }
     }
 }
