+++ Stormpath.SDK/Stormpath.SDK.Tests/Impl/Linq/ToListAsync_tests.cs
@@ -31,7 +31,7 @@ namespace Stormpath.SDK.Tests.Impl.Linq
     public class ToListAsync_tests : Linq_tests
     {
         [Fact]
-        public async Task ToListAsync_returns_empty_list_for_no_items()
+        public async Task Returns_empty_list_for_no_items()
         {
             var fakeDataStore = new FakeDataStore<IAccount>(Enumerable.Empty<IAccount>());
             var harness = CollectionTestHarness<IAccount>.Create<IAccount>(this.Href, fakeDataStore);
@@ -42,7 +42,7 @@ namespace Stormpath.SDK.Tests.Impl.Linq
         }
 
         [Fact]
-        public async Task ToListAsync_retrieves_all_items()
+        public async Task Retrieves_all_items()
         {
             var fakeDataStore = new FakeDataStore<IAccount>(FakeAccounts.RebelAlliance);
             var harness = CollectionTestHarness<IAccount>.Create<IAccount>(this.Href, fakeDataStore);
@@ -53,7 +53,7 @@ namespace Stormpath.SDK.Tests.Impl.Linq
         }
 
         [Fact]
-        public async Task ToListAsync_checks_for_new_items_after_last_page()
+        public async Task Checks_for_new_items_after_last_page()
         {
             // Scenario: 51 items in a server-side collection. The default limit is 25,
             // so two calls will return 25 items, and the 3rd will return 1. However, ToListAsync
@@ -66,100 +66,6 @@ namespace Stormpath.SDK.Tests.Impl.Linq
 
             longList.Count.ShouldBe(51);
             fakeDataStore.GetCalls().Count().ShouldBe(4);
-        }
-
-        [Fact]
-        public async Task ToListAsync_observes_take_limit()
-        {
-            // Scenario: .Take() functions a little differently than the limit=? parameter
-            // in Stormpath, even though that's what it translates to. .Take() represents an
-            // upper limit to the items that are returned. Take(5) returns 5 items, Take(500) returns 500.
-            var fakeDataStore = new FakeDataStore<IAccount>(Enumerable.Repeat(new FakeAccount(), 51));
-            var harness = CollectionTestHarness<IAccount>.Create<IAccount>(this.Href, fakeDataStore);
-
-            var longList = await harness.Queryable
-                .Take(7)
-                .ToListAsync();
-
-            longList.Count.ShouldBe(7);
-            fakeDataStore.GetCalls().Count().ShouldBe(1);
-        }
-
-        [Fact]
-        public async Task ToListAsync_pages_until_take_limit_is_reached()
-        {
-            // Scenario: .Take() functions a little differently than the limit=? parameter
-            // in Stormpath, even though that's what it translates to. .Take() represents an
-            // upper limit to the items that are returned. Take(5) returns 5 items, Take(500) returns 500.
-            var fakeDataStore = new FakeDataStore<IAccount>(Enumerable.Repeat(new FakeAccount(), 750));
-            var harness = CollectionTestHarness<IAccount>.Create<IAccount>(this.Href, fakeDataStore);
-
-            var longList = await harness.Queryable
-                .Take(500)
-                .ToListAsync();
-
-            longList.Count.ShouldBe(500);
-            fakeDataStore.GetCalls().Count().ShouldBe(5); // Max 100 per call
-        }
-
-        [Fact]
-        public async Task ForEachAsync_operates_on_every_item()
-        {
-            var harness = CollectionTestHarness<IAccount>.Create<IAccount>(
-                this.Href,
-                new FakeDataStore<IAccount>(FakeAccounts.RebelAlliance));
-            var gmailAlliance = new List<string>();
-
-            await harness.Queryable.ForEachAsync(acct =>
-            {
-                gmailAlliance.Add($"{acct.GivenName.ToLower()}@gmail.com");
-            });
-
-            gmailAlliance.Count.ShouldBe(FakeAccounts.RebelAlliance.Count);
-        }
-
-        [Fact]
-        public async Task ForEachAsync_indexes_every_item()
-        {
-            var harness = CollectionTestHarness<IAccount>.Create<IAccount>(
-                this.Href,
-                new FakeDataStore<IAccount>(FakeAccounts.GalacticEmpire));
-            var empireFirstNameLookup = new Dictionary<int, string>();
-
-            await harness.Queryable.ForEachAsync((acct, index) =>
-            {
-                empireFirstNameLookup.Add(index, $"{acct.GivenName} {acct.Surname}");
-            });
-
-            empireFirstNameLookup[2].ShouldBe(FakeAccounts.GalacticEmpire.ElementAt(2).GivenName + " " + FakeAccounts.GalacticEmpire.ElementAt(2).Surname);
-        }
-
-        [Fact]
-        public async Task ForEachAsync_can_be_cancelled()
-        {
-            var harness = CollectionTestHarness<IAccount>.Create<IAccount>(
-                this.Href,
-                new FakeDataStore<IAccount>(FakeAccounts.GalacticEmpire));
-            var cts = new CancellationTokenSource();
-            var reachedIndex = -1;
-
-            try
-            {
-                await harness.Queryable.ForEachAsync(
-                    (acct, index) =>
-                {
-                    reachedIndex = index;
-
-                    if (index == 2)
-                        cts.Cancel();
-                }, cts.Token);
-
-                Assertly.Fail("Should not reach here!");
-            }
-            catch (OperationCanceledException)
-            {
-                reachedIndex.ShouldBe(2);
-            }
         }
     }
 }
